Updating optimizations for speed
 * Adding support for trimmed NUFFT operations based on interpolator vector in Ggrid

 * Adding support in FieldCorrection to use trimmed operations

 * Changed default recons to float from double to make a better performance comparison with MATLAB

+++ PowerGrid/PowerGridDWI3D.cpp
@@ -73,7 +73,7 @@ int main(int argc, char** argv)
         }
 
         //int test = test_SpeedCompareGgrid<double>(testPath, Nx,Ny,Nz,NL,Niter,Ncoils);
-		int test = test_DWI<double>(testPath, Nx,Ny,Nz,NL,Niter,Ncoils);
+        int test = test_DWI<float>(testPath, Nx, Ny, Nz, NL, Niter, Ncoils);
 
     }
